Applied feedback to 404 page
